fix: refactor estimateChanges remove unused pages routes
