+++ src/components/permissions/Permissions.vue
@@ -50,7 +50,7 @@ function openLink() {
           {{ t("dappTransfer.permission") }}
         </p>
       </div>
-      <div class="mt-4 items-center px-4 flex flex-col justify-start items-start w-full">
+      <div class="mt-4 items-center px-4 flex flex-col justify-start w-full">
         <div class="flex flex-col justify-start items-start mt-4 mb-8 w-full">
           <p class="text-sm text-app-text-600 dark:text-app-text-dark-500">{{ `${t("dappInfo.requestFrom")}:` }}</p>
           <div class="w-full flex flex-row justify-between items-center bg-white dark:bg-app-gray-700 h-12 px-5 mt-3 rounded-md">
