+++ src/pages/redirectflow/Confirm.vue
@@ -4,17 +4,19 @@ import log from "loglevel";
 import { onErrorCaptured, onMounted, ref } from "vue";
 
 import { PaymentConfirm } from "@/components/payments";
+import { useEstimateChanges } from "@/components/payments/EstimateChangesComposable";
 import PermissionsTx from "@/components/permissionsTx/PermissionsTx.vue";
 import { TransactionChannelDataType } from "@/utils/enums";
 import { hideCrispButton, openCrispChat } from "@/utils/helpers";
 import { DecodedDataType, decodeInstruction } from "@/utils/instruction_decoder";
-import { AccountEstimation, FinalTxData } from "@/utils/interfaces";
+import { FinalTxData } from "@/utils/interfaces";
 import { redirectToResult, useRedirectFlow } from "@/utils/redirectflow_helpers";
-import { calculateTxFee, decodeAllInstruction, getEstimateBalanceChange, parsingTransferAmount } from "@/utils/solanaHelpers";
+import { calculateTxFee, decodeAllInstruction, parsingTransferAmount } from "@/utils/solanaHelpers";
 
 import ControllerModule from "../../modules/controllers";
 
 const { params, method, jsonrpc, req_id, resolveRoute } = useRedirectFlow();
+const { hasEstimationError, estimatedBalanceChange, estimationInProgress, estimateChanges } = useEstimateChanges();
 
 const finalTxData = ref<FinalTxData>();
 
@@ -22,22 +24,6 @@ const tx = ref<Transaction>();
 const decodedInst = ref<DecodedDataType[]>();
 const origin = ref("");
 const network = ref("");
-
-const hasEstimationError = ref("");
-const estimatedBalanceChange = ref<AccountEstimation[]>([]);
-const estimationInProgress = ref(true);
-
-const estimateChanges = async (transaction: Transaction) => {
-  estimationInProgress.value = true;
-  try {
-    estimatedBalanceChange.value = await getEstimateBalanceChange(ControllerModule.torus.connection, transaction, ControllerModule.selectedAddress);
-    hasEstimationError.value = "";
-  } catch (e) {
-    hasEstimationError.value = "Unable estimate balance changes";
-    log.info("Error in transaction simulation", e);
-  }
-  estimationInProgress.value = false;
-};
 
 onErrorCaptured(() => {
   openCrispChat();
@@ -74,7 +60,7 @@ onMounted(async () => {
       tx.value = Transaction.from(Buffer.from(txData.message as string, "hex"));
     }
 
-    estimateChanges(tx.value);
+    estimateChanges(tx.value, ControllerModule.connection, ControllerModule.selectedAddress);
     const isGasless = tx.value.feePayer?.toBase58() !== txData.signer;
     const txFee = isGasless ? 0 : (await calculateTxFee(tx.value.compileMessage(), ControllerModule.connection)).fee;
