+++ src/pages/redirectflow/ConfirmNft.vue
@@ -1,38 +1,22 @@
 <script setup lang="ts">
 import { LAMPORTS_PER_SOL, Transaction } from "@solana/web3.js";
-import log from "loglevel";
 import { computed, onMounted, ref, watch } from "vue";
 
+import { useEstimateChanges } from "@/components/payments/EstimateChangesComposable";
 import { getTokenFromMint, nftTokens } from "@/components/transfer/token-helper";
 import TransferNFT from "@/components/transfer/TransferNFT.vue";
-import { AccountEstimation } from "@/utils/interfaces";
 
 import ControllerModule from "../../modules/controllers";
 import { delay } from "../../utils/helpers";
 import { redirectToResult, useRedirectFlow } from "../../utils/redirectflow_helpers";
-import { calculateTxFee, generateSPLTransaction, getEstimateBalanceChange } from "../../utils/solanaHelpers";
+import { calculateTxFee, generateSPLTransaction } from "../../utils/solanaHelpers";
 
 const { params, method, resolveRoute, req_id, jsonrpc } = useRedirectFlow();
+const { hasEstimationError, estimatedBalanceChange, estimationInProgress, estimateChanges } = useEstimateChanges();
 
 const transactionFee = ref(0);
 const transaction = ref<Transaction>();
 const selectedNft = computed(() => getTokenFromMint(nftTokens.value, params.mint_add));
-
-const hasEstimationError = ref("");
-const estimatedBalanceChange = ref<AccountEstimation[]>([]);
-const estimationInProgress = ref(true);
-
-const estimateChanges = async (tx: Transaction) => {
-  estimationInProgress.value = true;
-  try {
-    estimatedBalanceChange.value = await getEstimateBalanceChange(ControllerModule.torus.connection, tx, ControllerModule.selectedAddress);
-    hasEstimationError.value = "";
-  } catch (e) {
-    hasEstimationError.value = "Unable estimate balance changes";
-    log.info("Error in transaction simulation", e);
-  }
-  estimationInProgress.value = false;
-};
 
 onMounted(async () => {
   // TODO: This can't be guaranteed
@@ -55,7 +39,7 @@ watch(selectedNft, async () => {
     );
 
     const { fee } = await calculateTxFee(transaction.value.compileMessage(), ControllerModule.connection);
-    estimateChanges(transaction.value);
+    estimateChanges(transaction.value, ControllerModule.connection, ControllerModule.selectedAddress);
     transactionFee.value = fee / LAMPORTS_PER_SOL;
   } else {
     redirectToResult(jsonrpc, { message: "Selected NFT not found", success: false, method }, req_id, resolveRoute);
