+++ src/pages/redirectflow/ConfirmSpl.vue
@@ -1,38 +1,23 @@
 <script setup lang="ts">
 import { LAMPORTS_PER_SOL, Transaction } from "@solana/web3.js";
-import log from "loglevel";
 import { computed, onMounted, ref, watch } from "vue";
 
+import { useEstimateChanges } from "@/components/payments/EstimateChangesComposable";
 import { tokens } from "@/components/transfer/token-helper";
 import TransferSPL from "@/components/transfer/TransferSPL.vue";
 import { delay } from "@/utils/helpers";
-import { AccountEstimation } from "@/utils/interfaces";
-import { calculateTxFee, generateSPLTransaction, getEstimateBalanceChange } from "@/utils/solanaHelpers";
+import { calculateTxFee, generateSPLTransaction } from "@/utils/solanaHelpers";
 
 import ControllerModule from "../../modules/controllers";
 import { redirectToResult, useRedirectFlow } from "../../utils/redirectflow_helpers";
 
 const { params, method, resolveRoute, jsonrpc, req_id } = useRedirectFlow();
+const { hasEstimationError, estimatedBalanceChange, estimationInProgress, estimateChanges } = useEstimateChanges();
 
 const transactionFee = ref(0);
 const transaction = ref<Transaction>();
 const selectedSplToken = computed(() => tokens.value.find((token) => token.mintAddress === params.mint_add));
 
-const hasEstimationError = ref("");
-const estimatedBalanceChange = ref<AccountEstimation[]>([]);
-const estimationInProgress = ref(true);
-
-const estimateChanges = async (tx: Transaction) => {
-  estimationInProgress.value = true;
-  try {
-    estimatedBalanceChange.value = await getEstimateBalanceChange(ControllerModule.torus.connection, tx, ControllerModule.selectedAddress);
-    hasEstimationError.value = "";
-  } catch (e) {
-    hasEstimationError.value = "Unable estimate balance changes";
-    log.info("Error in transaction simulation", e);
-  }
-  estimationInProgress.value = false;
-};
 onMounted(async () => {
   // This can't be guaranteed
   setTimeout(() => {
@@ -56,7 +41,7 @@ watch(selectedSplToken, async () => {
     );
 
     const { fee } = await calculateTxFee(transaction.value.compileMessage(), ControllerModule.connection);
-    estimateChanges(transaction.value);
+    estimateChanges(transaction.value, ControllerModule.connection, ControllerModule.selectedAddress);
     transactionFee.value = fee / LAMPORTS_PER_SOL;
   } else {
     redirectToResult(jsonrpc, { message: "Selected SPL token not found", success: false, method }, req_id, resolveRoute);
