+++ src/router.ts
@@ -130,18 +130,6 @@ const router = createRouter({
       meta: { title: "Confirm" },
     },
     {
-      name: "confirm_nft",
-      path: "/confirm_nft",
-      component: () => import(/* webpackPrefetch: true */ /* webpackChunkName: "CONFIRM_NFT" */ "@/pages/ConfirmNft.vue"),
-      meta: { title: "Confirm Nft" },
-    },
-    {
-      name: "confirm_spl",
-      path: "/confirm_spl",
-      component: () => import(/* webpackPrefetch: true */ /* webpackChunkName: "CONFIRM_SPL" */ "@/pages/ConfirmSpl.vue"),
-      meta: { title: "Confirm Spl" },
-    },
-    {
       name: "confirm_message",
       path: "/confirm_message",
       component: () => import(/* webpackPrefetch: true */ /* webpackChunkName: "CONFIRM_MESSAGE" */ "@/pages/ConfirmMessage.vue"),
